requirements.txt
kagglehub
numpy
pillow
matplotlib
scikit-learn
tqdm

torch
torchvision